requirements.txt
streamlit
streamlit-chat
watchdog
openai
python-dotenv